+++ packages/kit/src/core/config/index.spec.js
@@ -1,6 +1,6 @@
-import { suite, test } from 'uvu';
+import { test } from 'uvu';
 import * as assert from 'uvu/assert';
-import { deep_merge, validate_config } from './index.js';
+import { validate_config } from './index.js';
 
 test('fills in defaults', () => {
 	const validated = validate_config({});
@@ -9,7 +9,6 @@ test('fills in defaults', () => {
 	delete validated.kit.vite;
 
 	assert.equal(validated, {
-		compilerOptions: null,
 		extensions: ['.svelte'],
 		kit: {
 			adapter: null,
@@ -30,15 +29,15 @@ test('fills in defaults', () => {
 			i18n: null,
 			package: {
 				dir: 'package',
+				emitTypes: true,
 				exports: {
 					include: ['**'],
-					exclude: ['_*', '**/_*']
+					exclude: ['**/_*']
 				},
 				files: {
 					include: ['**'],
 					exclude: []
-				},
-				emitTypes: true
+				}
 			},
 			serviceWorker: {
 				exclude: []
@@ -50,17 +49,16 @@ test('fills in defaults', () => {
 			prerender: {
 				crawl: true,
 				enabled: true,
-				// TODO: remove this for the 1.0 release
+				entries: ['*'],
 				force: undefined,
 				onError: 'fail',
-				pages: ['*']
+				pages: undefined
 			},
 			router: true,
 			ssr: true,
 			target: null,
 			trailingSlash: 'never'
-		},
-		preprocess: null
+		}
 	});
 });
 
@@ -120,7 +118,6 @@ test('fills in partial blanks', () => {
 	delete validated.kit.vite;
 
 	assert.equal(validated, {
-		compilerOptions: null,
 		extensions: ['.svelte'],
 		kit: {
 			adapter: null,
@@ -141,15 +138,15 @@ test('fills in partial blanks', () => {
 			i18n: null,
 			package: {
 				dir: 'package',
+				emitTypes: true,
 				exports: {
 					include: ['**'],
-					exclude: ['_*', '**/_*']
+					exclude: ['**/_*']
 				},
 				files: {
 					include: ['**'],
 					exclude: []
-				},
-				emitTypes: true
+				}
 			},
 			serviceWorker: {
 				exclude: []
@@ -161,17 +158,16 @@ test('fills in partial blanks', () => {
 			prerender: {
 				crawl: true,
 				enabled: true,
-				// TODO: remove this for the 1.0 release
+				entries: ['*'],
 				force: undefined,
 				onError: 'fail',
-				pages: ['*']
+				pages: undefined
 			},
 			router: true,
 			ssr: true,
 			target: null,
 			trailingSlash: 'never'
-		},
-		preprocess: null
+		}
 	});
 });
 
@@ -192,7 +188,7 @@ test('fails if kit.appDir is only slash', () => {
 				appDir: '/'
 			}
 		});
-	}, /^kit\.appDir cannot start or end with '\/'. See https:\/\/kit\.svelte\.dev\/docs#configuration$/);
+	}, /^config\.kit\.appDir cannot start or end with '\/'. See https:\/\/kit\.svelte\.dev\/docs#configuration$/);
 });
 
 test('fails if kit.appDir starts with slash', () => {
@@ -202,7 +198,7 @@ test('fails if kit.appDir starts with slash', () => {
 				appDir: '/_app'
 			}
 		});
-	}, /^kit\.appDir cannot start or end with '\/'. See https:\/\/kit\.svelte\.dev\/docs#configuration$/);
+	}, /^config\.kit\.appDir cannot start or end with '\/'. See https:\/\/kit\.svelte\.dev\/docs#configuration$/);
 });
 
 test('fails if kit.appDir ends with slash', () => {
@@ -212,7 +208,7 @@ test('fails if kit.appDir ends with slash', () => {
 				appDir: '_app/'
 			}
 		});
-	}, /^kit\.appDir cannot start or end with '\/'. See https:\/\/kit\.svelte\.dev\/docs#configuration$/);
+	}, /^config\.kit\.appDir cannot start or end with '\/'. See https:\/\/kit\.svelte\.dev\/docs#configuration$/);
 });
 
 test('fails if paths.base is not root-relative', () => {
@@ -224,7 +220,7 @@ test('fails if paths.base is not root-relative', () => {
 				}
 			}
 		});
-	}, /^kit\.paths\.base option must be a root-relative path that starts but doesn't end with '\/'. See https:\/\/kit\.svelte\.dev\/docs#configuration-paths$/);
+	}, /^config\.kit\.paths\.base option must be a root-relative path that starts but doesn't end with '\/'. See https:\/\/kit\.svelte\.dev\/docs#configuration-paths$/);
 });
 
 test("fails if paths.base ends with '/'", () => {
@@ -236,7 +232,7 @@ test("fails if paths.base ends with '/'", () => {
 				}
 			}
 		});
-	}, /^kit\.paths\.base option must be a root-relative path that starts but doesn't end with '\/'. See https:\/\/kit\.svelte\.dev\/docs#configuration-paths$/);
+	}, /^config\.kit\.paths\.base option must be a root-relative path that starts but doesn't end with '\/'. See https:\/\/kit\.svelte\.dev\/docs#configuration-paths$/);
 });
 
 test('fails if paths.assets is relative', () => {
@@ -248,7 +244,7 @@ test('fails if paths.assets is relative', () => {
 				}
 			}
 		});
-	}, /^kit\.paths\.assets option must be an absolute path, if specified. See https:\/\/kit\.svelte\.dev\/docs#configuration-paths$/);
+	}, /^config\.kit\.paths\.assets option must be an absolute path, if specified. See https:\/\/kit\.svelte\.dev\/docs#configuration-paths$/);
 });
 
 test('fails if paths.assets has trailing slash', () => {
@@ -260,19 +256,19 @@ test('fails if paths.assets has trailing slash', () => {
 				}
 			}
 		});
-	}, /^kit\.paths\.assets option must not end with '\/'. See https:\/\/kit\.svelte\.dev\/docs#configuration-paths$/);
+	}, /^config\.kit\.paths\.assets option must not end with '\/'. See https:\/\/kit\.svelte\.dev\/docs#configuration-paths$/);
 });
 
-test('fails if prerender.pages are invalid', () => {
+test('fails if prerender.entries are invalid', () => {
 	assert.throws(() => {
 		validate_config({
 			kit: {
 				prerender: {
-					pages: ['foo']
+					entries: ['foo']
 				}
 			}
 		});
-	}, /^Each member of config\.kit.prerender.pages must be either '\*' or an absolute path beginning with '\/' — saw 'foo'$/);
+	}, /^Each member of config\.kit.prerender.entries must be either '\*' or an absolute path beginning with '\/' — saw 'foo'$/);
 });
 
 test('fails if i18n.defaultLocale is an empty string', () => {
@@ -354,179 +350,3 @@ validate_paths(
 );
 
 test.run();
-
-const deepMergeSuite = suite('deep_merge');
-
-deepMergeSuite('basic test no conflicts', async () => {
-	const [merged, conflicts] = deep_merge(
-		{
-			version: 1,
-			animalSounds: {
-				cow: 'moo'
-			}
-		},
-		{
-			animalSounds: {
-				duck: 'quack'
-			},
-			locale: 'en_US'
-		}
-	);
-	assert.equal(merged, {
-		version: 1,
-		locale: 'en_US',
-		animalSounds: {
-			cow: 'moo',
-			duck: 'quack'
-		}
-	});
-	assert.equal(conflicts, []);
-});
-
-deepMergeSuite('three way merge no conflicts', async () => {
-	const [merged, conflicts] = deep_merge(
-		{
-			animalSounds: {
-				cow: 'moo'
-			}
-		},
-		{
-			animalSounds: {
-				duck: 'quack'
-			}
-		},
-		{
-			animalSounds: {
-				dog: {
-					singular: 'bark',
-					plural: 'barks'
-				}
-			}
-		}
-	);
-	assert.equal(merged, {
-		animalSounds: {
-			cow: 'moo',
-			duck: 'quack',
-			dog: {
-				singular: 'bark',
-				plural: 'barks'
-			}
-		}
-	});
-	assert.equal(conflicts, []);
-});
-
-deepMergeSuite('merge with conflicts', async () => {
-	const [merged, conflicts] = deep_merge(
-		{
-			person: {
-				firstName: 'John',
-				lastName: 'Doe',
-				address: {
-					line1: '123 Main St',
-					city: 'Seattle',
-					state: 'WA'
-				}
-			}
-		},
-		{
-			person: {
-				middleInitial: 'Q',
-				address: '123 Main St, Seattle, WA'
-			}
-		}
-	);
-	assert.equal(merged, {
-		person: {
-			firstName: 'John',
-			middleInitial: 'Q',
-			lastName: 'Doe',
-			address: '123 Main St, Seattle, WA'
-		}
-	});
-	assert.equal(conflicts, ['person.address']);
-});
-
-deepMergeSuite('merge with arrays', async () => {
-	const [merged] = deep_merge(
-		{
-			paths: ['/foo', '/bar']
-		},
-		{
-			paths: ['/alpha', '/beta']
-		}
-	);
-	assert.equal(merged, {
-		paths: ['/foo', '/bar', '/alpha', '/beta']
-	});
-});
-
-deepMergeSuite('empty', async () => {
-	const [merged] = deep_merge();
-	assert.equal(merged, {});
-});
-
-deepMergeSuite('mutability safety', () => {
-	const input1 = {
-		person: {
-			firstName: 'John',
-			lastName: 'Doe',
-			address: {
-				line1: '123 Main St',
-				city: 'Seattle'
-			}
-		}
-	};
-	const input2 = {
-		person: {
-			middleInitial: 'L',
-			lastName: 'Smith',
-			address: {
-				state: 'WA'
-			}
-		}
-	};
-	const snapshot1 = JSON.stringify(input1);
-	const snapshot2 = JSON.stringify(input2);
-
-	const [merged] = deep_merge(input1, input2);
-
-	// Mess with the result
-	merged.person.middleInitial = 'Z';
-	merged.person.address.zipCode = '98103';
-	merged.person = {};
-
-	// Make sure nothing in the inputs changed
-	assert.snapshot(snapshot1, JSON.stringify(input1));
-	assert.snapshot(snapshot2, JSON.stringify(input2));
-});
-
-deepMergeSuite('merge buffer', () => {
-	const [merged, conflicts] = deep_merge(
-		{
-			x: Buffer.from('foo', 'utf-8')
-		},
-		{
-			y: 12345
-		}
-	);
-	assert.equal(Object.keys(merged), ['x', 'y']);
-	assert.equal(conflicts.length, 0);
-});
-
-deepMergeSuite('merge including toString', () => {
-	const [merged, conflicts] = deep_merge(
-		{
-			toString: () => '',
-			constructor: () => ''
-		},
-		{
-			y: 12345
-		}
-	);
-	assert.equal(conflicts.length, 0);
-	assert.equal(Object.keys(merged), ['toString', 'constructor', 'y']);
-});
-
-deepMergeSuite.run();