+++ packages/kit/src/core/create_manifest_data/index.spec.js
@@ -135,18 +135,20 @@ test('creates routes with layout', () => {
 	]);
 });
 
-test('encodes invalid characters', () => {
+test('encoding of characters', () => {
 	const { components, routes } = create('samples/encoding');
 
 	// had to remove ? and " because windows
 
 	// const quote = 'samples/encoding/".svelte';
 	const hash = 'samples/encoding/#.svelte';
+	const potato = 'samples/encoding/土豆.svelte';
 	// const question_mark = 'samples/encoding/?.svelte';
 
 	assert.equal(components, [
 		layout,
 		error,
+		potato,
 		// quote,
 		hash
 		// question_mark
@@ -155,6 +157,7 @@ test('encodes invalid characters', () => {
 	assert.equal(
 		routes.map((p) => p.pattern),
 		[
+			/^\/%E5%9C%9F%E8%B1%86\/?$/,
 			// /^\/%22\/?$/,
 			/^\/%23\/?$/
 			// /^\/%3F\/?$/