+++ packages/kit/types/endpoint.d.ts
@@ -1,18 +1,7 @@
 import { ServerRequest } from './hooks';
-import { MaybePromise, ResponseHeaders } from './helper';
+import { JSONString, MaybePromise, ResponseHeaders } from './helper';
 
-type ToJSON = { toJSON(...args: any[]): JSONValue };
-type JSONValue = Exclude<JSONResponse, ToJSON>;
-type JSONResponse =
-	| string
-	| number
-	| boolean
-	| null
-	| JSONResponse[]
-	| ToJSON
-	| { [key: string]: JSONResponse };
-
-type DefaultBody = JSONResponse | Uint8Array;
+type DefaultBody = JSONString | Uint8Array;
 
 export interface EndpointOutput<Body extends DefaultBody = DefaultBody> {
 	status?: number;